+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Add token authentication by specifying token in browser url
 - Fix compiled translations not embedding in releases
+- Events pushed by providers now take precedence over pulled ones.
 
 
 0.5.12 (2019-04-05)